+++ demo/app/Sharp/Posts/PostForm.php
@@ -30,8 +30,6 @@ class PostForm extends SharpForm
 {
     use WithSharpFormEloquentUpdater;
 
-    protected ?string $formValidatorClass = PostValidator::class;
-
     public function buildFormFields(FieldsContainer $formFields): void
     {
         $formFields
@@ -191,6 +189,20 @@ class PostForm extends SharpForm
             ->setCustomTransformer('cover', new SharpUploadModelFormAttributeTransformer())
             ->setCustomTransformer('attachments[document]', new SharpUploadModelFormAttributeTransformer())
             ->transform(Post::with('cover', 'attachments', 'categories')->findOrFail($id));
+    }
+
+    public function rules(): array
+    {
+        return [
+            'title.fr' => ['required', 'string', 'max:150'],
+            'title.en' => ['required', 'string', 'max:150'],
+            'content.text.fr' => ['required', 'string', 'max:2000'],
+            'content.text.en' => ['required', 'string', 'max:2000'],
+            'published_at' => ['required', 'date'],
+            'attachments.*.title' => ['required', 'string', 'max:50'],
+            'attachments.*.link_url' => ['required_if:attachments.*.is_link,true,1', 'nullable', 'url', 'max:150'],
+            'attachments.*.document' => ['required_if:attachments.*.is_link,false,0'],
+        ];
     }
 
     public function update($id, array $data)
