+++ demo/app/Sharp/Profile/ProfileSingleForm.php
@@ -15,8 +15,6 @@ class ProfileSingleForm extends SharpSingleForm
 {
     use WithSharpFormEloquentUpdater;
 
-    protected ?string $formValidatorClass = MyProfileValidator::class;
-
     public function buildFormFields(FieldsContainer $formFields): void
     {
         $formFields
@@ -67,6 +65,11 @@ class ProfileSingleForm extends SharpSingleForm
 
     protected function updateSingle(array $data)
     {
+        $this->validate(
+            $data,
+            ['name' => ['required', 'string', 'max:300']]
+        );
+
         $user = auth()->user();
 
         $this->save($user, $data);
